refactor(dashboard): 세션 칩 줄을 SessionFilterBar로 분리
세션 칩 목록과 열림 개수, 모바일 세션 토글 자리를 컴포넌트로 뺐습니다.
DashboardView에서 Chip import가 사라집니다.

열림 개수는 props로 받지 않고 컴포넌트가 직접 셉니다. 목록과 숫자를 따로
받으면 폴링으로 세션이 늘었을 때 둘이 어긋날 수 있습니다. ModerationQueue가
건수를 items.length로 그리는 것과 같은 이유입니다.

세션 목록 쿼리 자체는 DashboardView에 남깁니다. sessions 배열을 쓰는 곳이
칩 줄 말고도 세 군데(toMeta의 세션 제목 조회, selectedSession 찾기, early
return 두 개)라, 자식이 직접 쿼리하면 로딩·에러 처리가 갈라집니다.

Refs #205
